+++ desktop/js/options.js
@@ -1,24 +1,45 @@
-var commandeId;
 
 $("body").delegate("#bt_selectoutsideTemp", 'click', function() {
      jeedom.cmd.getSelectModal({cmd: {type: 'info', subtype: 'binary'}}, function(result) {
          $('#outsideTemp').value(result.human);
-		 commande = result.cmd.id;
+         $('#outsideTemp').attr('cmd',result.cmd.id);
      });
 });
 
 
 $('#btOptionsSave').off().on('click', function () {
- 
-  
-  	console.log(jeedom.config.save({
+    jeedom.config.save({
       plugin: "teleinfo",
-      configuration: {"outside_temp" : commande},
+      configuration: {"outside_temp" : $('#outsideTemp').attr('cmd')},
 		error: function (error) {
 		$('#div_OptionsAlert').showAlert({message: error.message, level: 'danger'});
 		},
 		success: function (data) {
-          console.log(data);
+            $('#div_OptionsAlert').showAlert({message: 'Sauvegarde effectuée', level: 'success'});
+		}
+    });
+
+console.log($("#indexConsoHP").val());
+    console.log(jeedom.config.save({
+      plugin: "teleinfo",
+      configuration: {"indexConsoHP" : $("#indexConsoHP").val()},
+		error: function (error) {
+		$('#div_OptionsAlert').showAlert({message: error.message, level: 'danger'});
+		},
+		success: function (data) {
+
 		}
     }));
+
+    console.log(jeedom.config.save({
+      plugin: "teleinfo",
+      configuration: {"indexConsoHC" : $("#indexConsoHC").val()},
+		error: function (error) {
+		$('#div_OptionsAlert').showAlert({message: error.message, level: 'danger'});
+		},
+		success: function (data) {
+
+		}
+    }));
+
 });