+++ desktop/js/teleinfo.js
@@ -213,8 +213,11 @@ function addCmdToTable(_cmd) {
     }else if(init(_cmd.configuration['type']) == 'health'){
         var tr = '';
     }
+    else if (init(_cmd.configuration['type']) == 'stat'){
+        var tr = '<tr class="cmd" data-cmd_id="' + init(_cmd.id) + '" style="background-color: aliceblue;">';
+    }
     else{
-        var tr = '<tr class="cmd" data-cmd_id="' + init(_cmd.id) + '">';
+        var tr = '<tr class="cmd" data-cmd_id="' + init(_cmd.id) + '" style="background-color: #f2fff2;">';
     }
     if(init(_cmd.configuration['type']) != 'health'){
         tr += '<td>';