+++ docs/fr_FR/changelog.md
@@ -2,11 +2,15 @@
 
 Pour toutes demandes : [https://github.com/NextDom/plugin-teleinfo/issues](https://github.com/NextDom/plugin-teleinfo/issues)
 
-### 4.2.3 (11-03-2019)
+### 4.2.3 (12-03-2019)
 
 * Globalisation des index
 * Correction changement niveau log du démon
 * Correction dépendances
+* Coloration du type de commande
+* Correction sur le panel
+* Ajout de la configuration des index a utiliser
+* Ajout librairie tagsinput
 
 ### 4.2.2 (08-03-2019)